check added

+++ hive_storage/lib/model/data_model.dart
@@ -9,7 +9,7 @@ class DataModel {
   String name;
 
   @HiveField(1)
-  double score;
+  int score;
 }
 
 List<DataModel> dataModel = [];
